hw2/requirements.txt
h5py==2.8.0
imageio==2.4.1
Markdown==3.0.1
matplotlib==3.0.2
scikit-image==0.14.1
Pillow
scikit-learn==0.20.1
scipy==1.1.0
tensorboard==1.14.0
tensorflow_gpu==1.14.0
jupyter
